+++ packages/typir/src/services/caching.ts
@@ -107,6 +107,7 @@ export class DefaultTypeRelationshipCaching<LanguageType = unknown> implements T
 export interface LanguageNodeInferenceCaching {
     cacheSet(languageNode: unknown, type: Type): void;
     cacheGet(languageNode: unknown): Type | undefined;
+    cacheClear(): void
     pendingSet(languageNode: unknown): void;
     pendingClear(languageNode: unknown): void;
     pendingGet(languageNode: unknown): boolean;
@@ -137,6 +138,10 @@ export class DefaultLanguageNodeInferenceCaching implements LanguageNodeInferenc
         } else {
             return this.cache.get(languageNode) as (Type | undefined);
         }
+    }
+
+    cacheClear(): void {
+        this.cache.clear();
     }
 
     pendingSet(languageNode: unknown): void {